requirements.txt
python-dotenv
requests
uuid
